requirements.txt
# BreastHealth Monitor - Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.2.0
numpy==1.26.3

# Telegram bot
aiogram==3.3.0

# LLM (optional)
openai==1.10.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
